Fixed bug where gallons column displayed sum used in mpg calculation rather than the current value.
This only showed for those that included previous non-fillup entries. See the note below.

Also fixed bug where '0' was evaluating as true for fillup when instantiating a mileageEntry object.

"MAJOR: There is an error with the way the mileageEntry calculates gallons. In the original table, gallons was only the number of gallons for a given fillup. This class automatically adds previous fillups. The way it does it is pretty cool, but it needs to be done in a separate operation from the one that pulls the gallons for the given entry. This I didn't fix yet, but look at the gallons for 3/3/08 in the csv compared to my official spreadsheet"

git-svn-id: svn://localhost/trunk@13 7eb0a0fd-2b4e-5d48-be3d-a80426e74fde